Vim: Explicitly set background color
This change has no effect on light-background terminals, but setting
"background=dark" is necessary to avoid a flash of a white background at
Vim startup on dark-background terminals. There's probably a bug in my
color scheme, but I haven't been able to track it down. I've looked at
the relevant Vim help text, the example color schemes, and the color
scheme README, and I've tried various experiments, but nothing I do
achieves my goal of getting the background setting corretly set to
"dark" before the scheme is loaded (or when "set &bg" is set near the
beginning of the scheme script). This hack is good enough for now.